flatten more
